+++ templates/android/app/src/main/java/__PACKAGE_PATH__/LynxApplication.kt
@@ -2,28 +2,40 @@ package __PACKAGE_NAME__
 
 import android.app.Application
 import com.lynx.tasm.LynxEnv
-import com.lynx.tasm.LynxEnvBuilder
+import com.facebook.drawee.backends.pipeline.Fresco
+import com.facebook.imagepipeline.core.ImagePipelineConfig
+import com.facebook.imagepipeline.memory.PoolConfig
+import com.facebook.imagepipeline.memory.PoolFactory
+import com.lynx.service.http.LynxHttpService
+import com.lynx.service.image.LynxImageService
+import com.lynx.service.log.LynxLogService
+import com.lynx.tasm.service.LynxServiceCenter
 
 class LynxApplication : Application() {
 
     override fun onCreate() {
         super.onCreate()
+        initLynxService()
         initLynxEnv()
     }
 
+    private fun initLynxService() {
+        // init Fresco which is needed by LynxImageService
+        val factory = PoolFactory(PoolConfig.newBuilder().build())
+        val builder = ImagePipelineConfig.newBuilder(applicationContext).setPoolFactory(factory)
+        Fresco.initialize(applicationContext, builder.build())
+
+        LynxServiceCenter.inst().registerService(LynxImageService.getInstance())
+        LynxServiceCenter.inst().registerService(LynxLogService)
+        LynxServiceCenter.inst().registerService(LynxHttpService)
+    }
+
     private fun initLynxEnv() {
-        try {
-            val builder = LynxEnvBuilder(this)
-            builder.setMediaResourceFetcher(LocalMediaFetcher())
-            LynxEnv.inst().init(builder)
-        } catch (e: Exception) {
-            e.printStackTrace()
-            // Fallback: basic initialization without media fetcher
-            try {
-                LynxEnv.inst().init(this, null, null, null)
-            } catch (e2: Exception) {
-                e2.printStackTrace()
-            }
-        }
+        LynxEnv.inst().init(
+            this,
+            null,
+            null,
+            null
+        )
     }
 }